requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pyaudio>=0.2.11
numpy>=1.24.0
sounddevice>=0.4.6
python-dotenv>=1.0.0
streamlit-webrtc>=0.47.0
av>=10.0.0
